:art: move `facades` tests to `services`
