+++ app/tests/services/issuer/test_acapy_issuer_v2.py
@@ -14,7 +14,7 @@ from aries_cloudcontroller import (
 from assertpy import assert_that
 from mockito import when
 
-from app.facades.issuer.acapy_issuer_v2 import IssuerV2
+from app.services.issuer.acapy_issuer_v2 import IssuerV2
 from app.models.issuer import Credential
 from app.tests.util.mock import to_async
